fix: changes POJOs, ClassX methods and class list to enable dynamic data
The issue was I was storing Class info in a string so if I update for example teacher name that won't reflect to the class which already has that teacher assigned. I dad to change all of this so it pulls the data from the table of those properties not from class table.